# back/api/v1/openapi.yaml
openapi: 3.0.3
...
info:
  title: cursed_matrix API
  version: 0.2.0
  description: |
    Every timestamp is ISO-8601 with an explicit UTC offset. Calendar days are
    YYYY-MM-DD strings computed in the user's timezone, never the server's.
    The server returns codes and parameters, never rendered prose: the client
    renders them in EN or RU.
servers:
  - url: /api/v1
...
paths:
...
  /tasks:
    get:
      tags: [tasks]
      summary: The board
      operationId: listTasks
      description: |
        Every task the filter matches, subtasks included, each with its tag
        names. Not paginated: the board and the graph need the whole working
        set, which the free plan caps at 35 active tasks. Completed tasks are
        never deleted, though, so a result is capped at 500 and says when it
        was cut.
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/StatusFilter" } }
        - { name: quadrants, in: query, schema: { type: array, items: { $ref: "#/components/schemas/Quadrant" } } }
        - { name: colors, in: query, schema: { type: array, items: { $ref: "#/components/schemas/TaskColor" } } }
        - { name: tags, in: query, schema: { type: array, items: { type: string } } }
        - { name: deadline, in: query, schema: { $ref: "#/components/schemas/DeadlineWindow" } }
        - { name: topology, in: query, schema: { $ref: "#/components/schemas/TopologyFilter" } }
        - { name: query, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { $ref: "#/components/schemas/SortField" } }
        - { name: direction, in: query, schema: { $ref: "#/components/schemas/SortDirection" } }
      responses:
        "200":
          description: The matching tasks.
          content:
            application/json:
              schema:
                type: object
                required: [tasks, links, truncated]
                properties:
                  tasks: { type: array, items: { $ref: "#/components/schemas/Task" } }
                  links:
                    type: array
                    description: |
                      Every link the user holds. They travel with the board
                      because a card shows a link count and the topology filter
                      needs them; fetched separately the two answers could
                      disagree.
                    items: { $ref: "#/components/schemas/TaskLink" }
                  truncated:
                    type: boolean
                    description: |
                      True when the result hit the server cap of 500 tasks and
                      more exist. Narrow the filter to see the rest.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    post:
      tags: [tasks]
      summary: Create a task
      operationId: createTask
      description: |
        The task lands at the end of the quadrant it was created in. Position
        is the server's to decide; the client cannot ask for one.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskDraft" }
      responses:
        "201":
          description: The task.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/QuotaReached" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /tasks/{taskId}:
    parameters:
      - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [tasks]
      summary: Edit a task
      operationId: updateTask
      description: |
        Only the fields the body mentions change. `deadlineAt` sent as null
        clears the deadline; left out, it stays. Quadrant and position are not
        editable here — moving a task is its own operation, because it changes
        two things at once.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskPatch" }
      responses:
        "200":
          description: The task as it now stands.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The task is completed, and a completed task is frozen.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [tasks]
      summary: Delete a task
      operationId: deleteTask
      description: |
        A logical delete. The row stays so the XP ledger and the activity
        history keep referring to something real; the task leaves the user's
        views, and any XP it was paid is withdrawn by a compensating entry.
        Completed subtasks go with their parent; unfinished ones block the
        request.
      responses:
        "200":
          description: What was removed.
          content:
            application/json:
              schema:
                type: object
                required: [deletedIds]
                properties:
                  deletedIds:
                    type: array
                    items: { type: string, format: uuid }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            The task still has unfinished subtasks. Finish, promote or delete
            them first — they are open work, and the server will not take them
            down on the client's behalf.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /tasks/{taskId}/move:
    parameters:
      - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [tasks]
      summary: Move a task
      operationId: moveTask
      description: |
        The client names a place, never a position: two clients dragging at
        once would compute the same number from different starting points and
        the second write would land in the wrong place. Give `beforeTaskId` or
        `afterTaskId`, or neither to append. The server assigns the position
        and renumbers the quadrant when a gap is used up. Moving a subtask out
        with this call promotes it.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MoveRequest" }
      responses:
        "200":
          description: Every task whose position changed, the moved one last.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            The task is completed. The quadrant it was completed in is what
            priced its XP and is frozen with the snapshot; reopen it first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /tasks/{taskId}/promote:
    parameters:
      - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [tasks]
      summary: Turn a subtask into a task
      operationId: promoteTask
      description: |
        The parent relation is dropped and the task takes a place at the end of
        a quadrant — its former parent's unless the body names another. Tags,
        colour, deadline and the XP snapshot survive, so a completed subtask
        can be promoted without disturbing what it earned.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                quadrant: { $ref: "#/components/schemas/Quadrant" }
      responses:
        "200":
          description: Every task whose position changed, the promoted one last.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: The task has no parent, so there is nothing to promote.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /tasks/{taskId}/complete:
    parameters:
      - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [tasks]
      summary: Complete a task
      operationId: completeTask
      description: |
        Marks the task done and, for a parent, every active subtask under it
        with `completedVia: PARENT_CASCADE` and the same `completedAt`. XP is
        priced from the effective quadrant — a subtask's is its parent's — and
        frozen on the task, so a later change to the configuration never
        rewrites what was earned. Status, snapshot, ledger entry and totals are
        one transaction.
      responses:
        "200":
          description: Everything the completion changed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Progress" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            Already completed, or a subtask whose parent cannot be read, which
            must not be priced at a guessed quadrant.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /tasks/{taskId}/reopen:
    parameters:
      - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [tasks]
      summary: Return a task to work
      operationId: reopenTask
      description: |
        Clears the completion and its snapshot, and withdraws the XP with a
        compensating negative entry; the original entry is never deleted, so
        the history stays auditable. Subtasks completed by a cascade are not
        reopened.
      responses:
        "200":
          description: The task, with the withdrawn XP as a negative amount.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Progress" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /tasks/{taskId}/subtasks:
    parameters:
      - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [tasks]
      summary: Add a subtask
      operationId: createSubtask
      description: |
        A subtask has no quadrant of its own — it answers to its parent's — and
        cannot have subtasks itself. Subtasks count against the plan's task
        quota like any other task.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubtaskDraft" }
      responses:
        "201":
          description: The subtask.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/QuotaReached" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: cm_access
      description: |
        A signed JWT the page cannot read. Verified by signature, so an
        authenticated request touches neither PostgreSQL nor Redis.
    refreshCookie:
      type: apiKey
      in: cookie
      name: cm_refresh
      description: |
        The opaque id of a server-side record, scoped to /api/v1/auth and
        SameSite=Strict. Revocable, which is what the JWT cannot be.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        The value of the page-readable cm_csrf cookie. Required on every unsafe
        method, because a browser attaches cookies whether or not the page meant
        to send them.
...
  responses:
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    CSRFRejected:
      description: The CSRF header is missing or does not match the cookie.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationFailed:
      description: Valid JSON, invalid content.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: |
        No such task for this account. A task belonging to someone else answers
        the same way as one that does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    QuotaReached:
      description: |
        The plan's limit is reached. Nothing existing is affected; completing
        or deleting a task frees a slot.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code]
          properties:
            code:
              type: string
              description: |
                Stable and language-independent. The client renders it; the
                server never sends prose.
              example: VALIDATION_FAILED
            params:
              type: object
              additionalProperties: true
              description: Values for the client's message template.
            requestId:
              type: string
              description: What the user reports when something goes wrong.
...
    Quadrant:
      type: string
      enum: [IMPORTANT_URGENT, IMPORTANT_NOT_URGENT, NOT_IMPORTANT_URGENT, NOT_IMPORTANT_NOT_URGENT]
    TaskColor:
      type: string
      enum: [NONE, CYAN, VIOLET, AMBER, ROSE, TEAL, SLATE]
    TaskStatus:
      type: string
      enum: [ACTIVE, COMPLETED]
    CompletionSource:
      type: string
      enum: [DIRECT, PARENT_CASCADE]
    StatusFilter:
      type: string
      enum: [ACTIVE, COMPLETED, ALL]
    DeadlineWindow:
      type: string
      enum: [ANY, OVERDUE, TODAY, WEEK, NONE]
    TopologyFilter:
      type: string
      enum: [ANY, LINKED, UNLINKED]
    SortField:
      type: string
      enum: [POSITION, DEADLINE, CREATED, TITLE]
    SortDirection:
      type: string
      enum: [ASC, DESC]
...
    LinkType:
      type: string
      enum: [RELATED, CONNECTED_TO, BLOCKS, DEPENDS_ON]
...
    TaskLink:
      type: object
      required: [id, sourceTaskId, targetTaskId, type, createdAt]
      properties:
        id: { type: string, format: uuid }
        sourceTaskId: { type: string, format: uuid }
        targetTaskId: { type: string, format: uuid }
        type: { $ref: "#/components/schemas/LinkType" }
        createdAt: { type: string, format: date-time }
    MoveRequest:
      type: object
      required: [targetQuadrant]
      description: At most one neighbour; neither means the end of the list.
      properties:
        targetQuadrant: { $ref: "#/components/schemas/Quadrant" }
        beforeTaskId: { type: string, format: uuid }
        afterTaskId: { type: string, format: uuid }
    TaskList:
      type: object
      required: [tasks]
      properties:
        tasks: { type: array, items: { $ref: "#/components/schemas/Task" } }
    Progress:
      type: object
      description: |
        What one call changed. The tasks are returned so the client can
        reconcile its board without fetching it again.
      required: [tasks, xpAwarded, unlockedAchievements]
      properties:
        tasks: { type: array, items: { $ref: "#/components/schemas/Task" } }
        xpAwarded:
          type: integer
          format: int32
          description: Negative when XP was withdrawn.
        levelUp:
          type: object
          nullable: true
          description: One object however many thresholds were crossed at once.
          required: [fromLevel, toLevel]
          properties:
            fromLevel: { type: integer, format: int32 }
            toLevel: { type: integer, format: int32 }
        unlockedAchievements:
          type: array
          description: Achievement codes. Always empty until the evaluator exists.
          items: { type: string }
    TaskDraft:
      type: object
      required: [title, quadrant]
      properties:
        title: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 2000 }
        quadrant: { $ref: "#/components/schemas/Quadrant" }
        color: { $ref: "#/components/schemas/TaskColor" }
        deadlineAt: { type: string, format: date-time, nullable: true }
        deadlineHasTime: { type: boolean }
    SubtaskDraft:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 2000 }
        color: { $ref: "#/components/schemas/TaskColor" }
        deadlineAt: { type: string, format: date-time, nullable: true }
        deadlineHasTime: { type: boolean }
    TaskPatch:
      type: object
      description: Every field is optional; only what is present changes.
      properties:
        title: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 2000 }
        color: { $ref: "#/components/schemas/TaskColor" }
        deadlineAt:
          type: string
          format: date-time
          nullable: true
          description: Null clears the deadline; absent leaves it unchanged.
        deadlineHasTime: { type: boolean }
    Task:
      type: object
      description: |
        A regular task or a subtask; parentTaskId tells them apart. quadrant is
        non-null exactly when parentTaskId is null: a subtask has none of its
        own and inherits the parent's. The four completion fields are set
        together, exactly when status is COMPLETED, and are frozen at that
        moment so a later change to the XP configuration never rewrites
        history.
      required: [id, userId, title, description, position, color, deadlineHasTime, status, createdAt, updatedAt, tags]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        parentTaskId: { type: string, format: uuid, nullable: true }
        title: { type: string, maxLength: 100 }
        description: { type: string, maxLength: 2000 }
        quadrant: { allOf: [{ $ref: "#/components/schemas/Quadrant" }], nullable: true }
        position: { type: integer, format: int32 }
        color: { $ref: "#/components/schemas/TaskColor" }
        deadlineAt: { type: string, format: date-time, nullable: true }
        deadlineHasTime:
          type: boolean
          description: False means a date-only deadline, rendered without a time.
        status: { $ref: "#/components/schemas/TaskStatus" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
        xpAwarded: { type: integer, format: int32, nullable: true }
        quadrantAtCompletion: { allOf: [{ $ref: "#/components/schemas/Quadrant" }], nullable: true }
        completedVia: { allOf: [{ $ref: "#/components/schemas/CompletionSource" }], nullable: true }
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: |
            Always absent on the board, which never shows a deleted task. It
            is here for the export, which includes them — a list that carried
            them without saying which would be worse than leaving them out.
        tags: { type: array, items: { type: string } }
...
security:
  - accessCookie: []
    csrfToken: []